# thetagateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for ThetaPay backend.",
        "title": "ThetaPay API Docs",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/customers/login": {
            "post": {
                "description": "Cutomer Login with Email \u0026 Password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Customer Login",
                "parameters": [
                    {
                        "description": "Login Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Customer"
                        }
                    }
                }
            }
        },
        "/customers/signup": {
            "post": {
                "description": "Signup a new Customer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Customer SignUp",
                "parameters": [
                    {
                        "description": "Signup Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Customer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "description": "Users Login with Email \u0026 Password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Users Login",
                "parameters": [
                    {
                        "description": "Login Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AccountInfo": {
            "type": "object",
            "required": [
                "accountId",
                "accountType",
                "balance",
                "pincode"
            ],
            "properties": {
                "accountId": {
                    "type": "string"
                },
                "accountType": {
                    "type": "string"
                },
                "balance": {
                    "type": "number"
                },
                "pincode": {
                    "type": "integer"
                }
            }
        },
        "models.Customer": {
            "type": "object",
            "required": [
                "accountInfo",
                "createdAt",
                "email",
                "secretKey",
                "updatedAt"
            ],
            "properties": {
                "accountInfo": {
                    "$ref": "#/definitions/models.AccountInfo"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "secretKey": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "required": [
                "createdAt",
                "phoneNumber",
                "role",
                "updatedAt",
                "userName"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "role": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}